Fix editing posts which result in the word-count dropping under the threshold to memorize

+++ upload/src/addons/SV/WordCountSearch/XF/Entity/Post.php
@@ -147,7 +147,11 @@ class Post extends XFCP_Post
         else if ($this->Words)
         {
             $changes = true;
-            $this->Words->delete();
+            if ($this->Words->exists())
+            {
+                $this->Words->reset();
+                $this->Words->delete();
+            }
             $this->clearCache('Words');
         }
         $this->clearCache('WordCount');
